-notify cli parameter added.

+++ make_report.py
@@ -4,6 +4,8 @@ commits = {}
 personTemplate = GetTemplate("person")
 teams = yaml.load(ReadFile("conf/teams.yaml"))
 
+notify = "-notify" in sys.argv
+
 ######################################################################################
 # Git commits
 
@@ -18,24 +20,23 @@ text = GetStdoutOf("gitlog.bat", "%s %s" % (lastWorkday.strftime("%Y-%m-%d"), to
 # Split into lines and treat each
 [AddCommit(line, commits) for line in text.split("\n")]
 
-
 ######################################################################################
 # Jira worklogs
 
 print "\n-- Fetching jira worklogs: ---------------------------------------------"
 workLogs = GetWorkLogs(lastWorkday, today) 
 
-# Notify guys who forgot to fill worklog via give engine
-print "\n-- Sending IM notifications: -------------------------------------------"
-RequestWorklogs(lastWorkday, workLogs, config["notified_skype"], Skype(), commits)
+if notify:
+	# Notify guys who forgot to fill worklog via give engine
+	print "\n-- Sending IM notifications: -------------------------------------------"
+	RequestWorklogs(lastWorkday, workLogs, config["notified_skype"], Skype(), commits)
+else:
+	# Populate template with received values
+	page = FillTemplate(GetTemplate("report"), {"##SARATOV##": BindTeamLogs("Saratov", teams, commits, workLogs, personTemplate), "##US##": BindTeamLogs("US", teams, commits, workLogs, personTemplate), "##TODAY##": today.strftime("%Y-%m-%d")})
 
-# Populate template with received values
-page = FillTemplate(GetTemplate("report"), {"##SARATOV##": BindTeamLogs("Saratov", teams, commits, workLogs, personTemplate), "##US##": BindTeamLogs("US", teams, commits, workLogs, personTemplate), "##TODAY##": today.strftime("%Y-%m-%d")})
-
-
-WriteFile("temp1.tmp", page)
-#GetWiki({"action": "storePage", "space": "~nbogdanov", "title": "gitlog + %s report template" % today, "file": "temp1.tmp", "parent": "BigRock Reporting automation"})
-GetWiki({"action": "storeNews", "space": "ToRS", "title": "%s Daily Report" % today.strftime("%Y.%m.%d"), "file": "temp1.tmp"})
-os.remove("temp1.tmp")
+	WriteFile("temp1.tmp", page)
+	#GetWiki({"action": "storePage", "space": "~nbogdanov", "title": "gitlog + %s report template" % today, "file": "temp1.tmp", "parent": "BigRock Reporting automation"})
+	GetWiki({"action": "storeNews", "space": "ToRS", "title": "%s Daily Report" % today.strftime("%Y.%m.%d"), "file": "temp1.tmp"})
+	os.remove("temp1.tmp")
 
 print "Done"
